Add design document for launch (#184)
I'm planning to refactor some of the code to support the ideas described
in this document. I thought I'd split this out into a PR so it can be
reviewed and shot at earlier. I realize that if we merge this, the
implementation and documentation are out of sync. I view this more as a
target that the implementation should aim to follow.