Mick/make databricks config optional (#208)
I'm not sure if the change in this PR is what we actually want. 

Can we think about whether we:

- want to upload the databricks secret from user's machines at all, or
just have a secret preconfigured in the cluster that everyone uses
- should implement various modifications to the job/rayjobs (like the
databrickscfg volume) inside of the launch tool, or instead allow
passing a yml template file ([like
MLflow](https://github.com/mlflow/mlflow/blob/master/examples/docker/kubernetes_job_template.yaml)),
or both?

If we need user-specific secrets, we should probably think about how to
make those secrets in the cluster only usable by the user that created
them. The current implementation uses a global secret named
`databrickcfg`. There is a race condition since secret (re)creation and
usage in a job are unsynchronized. In the general case it is also a
security concern where users can access each other's secrets.

The second point is less pressing. We will find out how much
specialization we need for different projects and how we can support it.

If we don't want per-user databricks secrets, we still must decide if we
want to 1) always provide the databricks volume, 2) require passing a
flag to provide the volume or do something else. Probably 1) is good
enough.

---------

Co-authored-by: Gary Miguel <garymm@astera.org>