Update launch from source installation instructions (#763)
This is an easier alternative but probably with worse caching (not sure
if cargo install caches). Obviously slower than installing pre-built
binaries. This requires `cargo` to be available which can be installed
with `rustup`, but does not require `pixi` to be available which we
intend to phase out and it does not require us to set up binary build
infrastructure (for now).