Implement automatic update warning for launch (#835)
Fixes #377 

The version check does not block process exit and executes on regular
termination and SIGINT termination (for long running commands).

This check will always execute but may not finish and should not crash
the process. It will also execute for local builds and irrespective of
how launch was installed.