Check if image exists, before building (#838)
Launch always builds an image, before running it. With kaniko, this is
slower since we don't have caching.
So we can tag the docker images using the git commit and check if the
image exists, before building it un-necessarily.
This will allow repeated launch runs of the same commit, but with
different launch args, to be much faster.

Helps toward #804

---------

Co-authored-by: Mick van Gelderen <mickvangelderen@gmail.com>